feat : SecurityContextInterceptor 추상화 구현
